pyproject.toml
[tool.poetry]
name = "cch-todo"
version = "0.1.0"
description = ""
authors = ["cch0807 <ckdgus11021@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "cch_todo"}]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.23.2"
fastapi = "0.97.0"
black = "^23.7.0"
flake8 = "^6.1.0"
sqlalchemy = "^2.0.19"
alembic = "^1.11.2"
pymysql = "^1.1.0"
cryptography = "^41.0.3"
pytz = "^2023.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
